feat: CalculatorItem 케이스 수정
