+++ Calculator/Calculator/Model/Assets.xcassets/AccentColor.colorset/Contents.json
