+++ Calculator/Calculator/Model/Assets.xcassets/Contents.json
